Added damage variable to Bullet to allow weapon upgrades Added more variables affected by difficulty selection: Mob-Health, Max-Mobs, Mob-Strength Changed health to float for more flexibility

+++ src/com/mojang/mojam/entity/Player.java
@@ -449,7 +449,7 @@ public class Player extends Mob implements LootCollector {
 	}
 
 	@Override
-	public void hurt(Entity source, int damage) {
+	public void hurt(Entity source, float damage) {
 		if (isImmortal) {
 			return;
 		}
